Ignore USB mass storage devices in LIVE mode
This avoids a 30s lock-up at boot, due to system_server scanning the USB
bus, interfering with the live system that is booted from the USB drive.
This usually results in a ANR dialog for SystemUI, and in some rare
occasions, a full reboot of the device.

This change uses the ro.boot.live property to determine whether USB
mass storage devices should be ignored by UsbHostManager.

Change-Id: I540a985e6f48c534988791874088b4967aac4a09